refactor(entity): move helpers to utils
